+++ api/spec/controllers/spree/api/return_authorizations_controller_spec.rb
@@ -105,6 +105,53 @@ module Spree
         json_response.should have_attributes(attributes)
       end
 
+      it "can add an inventory unit to a return authorization on the order" do
+        FactoryGirl.create(:return_authorization, :order => order)
+        return_authorization = order.return_authorizations.first
+        inventory_unit = return_authorization.returnable_inventory.first
+        inventory_unit.should be
+        return_authorization.inventory_units.should be_empty
+        api_put :add, :id => return_authorization.id, variant_id: inventory_unit.variant.id, quantity: 1
+        response.status.should == 200
+        json_response.should have_attributes(attributes)
+        return_authorization.reload.inventory_units.should_not be_empty
+      end
+
+      it "can mark a return authorization as received on the order with an inventory unit" do
+        FactoryGirl.create(:new_return_authorization, :order => order)
+        return_authorization = order.return_authorizations.first
+        return_authorization.state.should == "authorized"
+
+        # prep (use a rspec context or a factory instead?)
+        inventory_unit = return_authorization.returnable_inventory.first
+        inventory_unit.should be
+        return_authorization.inventory_units.should be_empty
+        api_put :add, :id => return_authorization.id, variant_id: inventory_unit.variant.id, quantity: 1
+        # end prep
+
+        api_delete :receive, :id => return_authorization.id
+        response.status.should == 200
+        return_authorization.reload.state.should == "received"
+      end
+
+      it "cannot mark a return authorization as received on the order with no inventory units" do
+        FactoryGirl.create(:new_return_authorization, :order => order)
+        return_authorization = order.return_authorizations.first
+        return_authorization.state.should == "authorized"
+        api_delete :receive, :id => return_authorization.id
+        response.status.should == 422
+        return_authorization.reload.state.should == "authorized"
+      end
+
+      it "can cancel a return authorization on the order" do
+        FactoryGirl.create(:new_return_authorization, :order => order)
+        return_authorization = order.return_authorizations.first
+        return_authorization.state.should == "authorized"
+        api_delete :cancel, :id => return_authorization.id
+        response.status.should == 200
+        return_authorization.reload.state.should == "canceled"
+      end
+
       it "can delete a return authorization on the order" do
         FactoryGirl.create(:return_authorization, :order => order)
         return_authorization = order.return_authorizations.first