+++ core/CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Spree 2.1.0 (unreleased) ##
 
+*   Change `order.promotion_credit_exists?` api. Now it receives an adjustment
+    originator (PromotionAction instance) instead of a promotion. Allowing
+    multiple adjustments being created for the same promotion as the current
+    PromotionAction / Promotion api suggests #3262
+
 *   Remove after_save callback for stock items backorders processing and
     fixes count on hand updates when there are backordered units #3066
 
@@ -52,3 +57,11 @@ en:
 See #3133 for more information.
 
     * Ryan Bigg*
+
+* CreditCard model now validates that the card is not expired.
+
+    *Ryan Bigg*
+
+* Payment model will now no longer provide a vague error message for when the source is invalid. Instead, it will provide error messages like "Credit Card Number can't be blank"
+
+    *Ryan Bigg*